+++ db/seeds.rb
@@ -57,6 +57,7 @@ class SeedDB
     call_and_log_seed_step Seeds::TestCaseData
     call_and_log_seed_step Seeds::Notifications
     call_and_log_seed_step Seeds::CavcDashboardData
+    call_and_log_seed_step Seeds::AdditionalRemandedAppeals
     # Always run this as last one
     call_and_log_seed_step Seeds::StaticTestCaseData
     call_and_log_seed_step Seeds::StaticDispatchedAppealsTestData
